test: add tests for celery app
