+++ tests/resources/test.conf
@@ -9,7 +9,7 @@ repo-config-email=test_email@example.com
 
 [Secrets-Section]
 flask-secret-key=S3CR3T
-rabbitmq-password=guest
+rabbitmq-password=rabbitmq
 opensearch-secret=test
 confd-credentials='test test'
 yang-catalog-token=test
@@ -75,3 +75,9 @@ non-ietf-directory=/var/yang/nonietf
 email-from=test
 email-to=test
 developers-email=test
+
+[RabbitMQ-Section]
+host=yc-rabbit
+username=rabbitmq
+virtual-host=/
+port=5672
